Refactor as ASGI app, move example scripts to root dir
