--- example/scripts/install
@@ -1,8 +1,0 @@
-#!/bin/sh -ex
-
-python -m venv venv
-
-venv/bin/pip install -U pip
-venv/bin/pip install -r requirements.txt
-
-echo "DEBUG=true" > .env
