--- example/scripts/serve
@@ -1,3 +1,0 @@
-#!/bin/sh -e
-
-exec venv/bin/uvicorn server:app
