+++ example/server/templates/index.jinja
@@ -9,7 +9,7 @@
     {{ page_content | safe }}
 
     {% if DEBUG %}
-      {{ hotreload.script(url_for('hot-reload')) | safe }}
+      {{ hotreload.script(url=url_for("hot-reload")) | safe }}
     {% endif %}
   </body>
 </html>
