+++ src/arel/__init__.py
@@ -1,4 +1,4 @@
 from .__version__ import __version__
-from ._main import HotReload
+from ._app import HotReload
 
 __all__ = ["__version__", "HotReload"]
